MR-182: member profile changed, new Profile shared component
